Make find_section_in_bytes test-only
